refactor(core): split bind validation, boxplot body, and finalize phases
Extract rule-form and geom contracts from bind-layer-validate; split boxplot
body vs outliers; and divide finalize into layout/geometry and model assembly.
Characterization tests pin boxplot outliers, continuous-x boxplot rejection,
and identity errorbar channel requirements.
